fix: state 및 컴포넌트 정리
