+++ src/features/users/components/LoginForm.tsx
@@ -1,8 +1,13 @@
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
-import type { UseFormHandleSubmit, UseFormRegister } from 'react-hook-form';
+import type {
+  UseFormHandleSubmit,
+  UseFormRegister,
+  UseFormWatch,
+} from 'react-hook-form';
 import type { userData } from '@/types/types';
+import { useState } from 'react';
 
 // const data = {
 //   id: 'gildongmom',
@@ -12,16 +17,19 @@ import type { userData } from '@/types/types';
 // };
 
 type LoginFormProps = {
+  watch: UseFormWatch<userData>;
   handleSubmit: UseFormHandleSubmit<userData>;
   register: UseFormRegister<userData>;
   setIsUser: React.Dispatch<React.SetStateAction<boolean>>;
 };
 
 export default function LoginForm({
+  watch,
   handleSubmit,
   register,
   setIsUser,
 }: LoginFormProps) {
+  const [isActive, setIsActive] = useState<boolean>(false);
   const LoginHandler = (
     setIsUser: React.Dispatch<React.SetStateAction<boolean>>,
     data: userData
@@ -29,9 +37,13 @@ export default function LoginForm({
     console.log(data);
     // 만약 기존 회원이 아닐 경우 isUser = false;
     // 잘못 입력했을 경우를 대비하여 confirm으로 확인.
-    if (confirm('잘못 입력된 정보입니다. 회원가입 하시겠습니까?'))
+    if (isActive && confirm('잘못 입력된 정보입니다. 회원가입 하시겠습니까?'))
       setIsUser(false);
   };
+
+  const idValue = watch('id');
+  const passwordValue = watch('password');
+
   return (
     <form
       id='login'
@@ -45,7 +57,10 @@ export default function LoginForm({
         id='id'
         placeholder='이름을 입력하세요.'
         className='px-[15px] py-[9px] h-[52px] box-border mb-2 border-primary border-[1px] rounded-[12px]'
-        {...register('id', { required: 'ID 입력은 필수입니다.' })}
+        {...register('id')}
+        onBlur={() => {
+          if (idValue !== '' && passwordValue !== '') setIsActive(true);
+        }}
       />
       <Label htmlFor='password' className='font-korean-title text-xl font-bold'>
         비밀번호
@@ -55,11 +70,16 @@ export default function LoginForm({
         id='password'
         placeholder='******'
         className='px-[15px] py-[9px] h-[52px] box-border mb-2 border-primary border-[1px] rounded-[12px]'
-        {...register('password', {
-          required: '비밀번호 입력은 필수입니다.',
-        })}
+        {...register('password')}
+        onBlur={() => {
+          if (idValue !== '' && passwordValue !== '') setIsActive(true);
+        }}
       />
-      <Button type='submit' className='h-[48px]'>
+      <Button
+        variant={`${isActive ? 'default' : 'disabled'}`}
+        type='submit'
+        className='h-[48px]'
+      >
         로그인
       </Button>
     </form>
